depricated: customer and payload class
